+++ packages/core/src/EntityManager.ts
@@ -1,7 +1,7 @@
 import { inspect } from 'util';
 import type { Configuration } from './utils';
 import { Cursor, QueryHelper, TransactionContext, Utils } from './utils';
-import type { AssignOptions, EntityLoaderOptions, EntityRepository, IdentifiedReference } from './entity';
+import type { AssignOptions, EntityLoaderOptions, EntityRepository } from './entity';
 import { EntityAssigner, EntityFactory, EntityLoader, EntityValidator, helper, Reference } from './entity';
 import { ChangeSet, ChangeSetType, UnitOfWork } from './unit-of-work';
 import type {
@@ -37,6 +37,7 @@ import type {
   PopulateOptions,
   Primary,
   RequiredEntityData,
+  Ref,
 } from './typings';
 import type { TransactionOptions } from './enums';
 import { FlushMode, LoadStrategy, LockMode, PopulateHint, ReferenceType, SCALAR_TYPES } from './enums';
@@ -1159,7 +1160,7 @@ export class EntityManager<D extends IDatabaseDriver = IDatabaseDriver> {
   /**
    * Gets a reference to the entity identified by the given type and identifier without actually loading it, if the entity is not yet loaded
    */
-  getReference<Entity extends object, PK extends keyof Entity>(entityName: EntityName<Entity>, id: Primary<Entity>, options: Omit<GetReferenceOptions, 'wrapped'> & { wrapped: true }): IdentifiedReference<Entity, PK>;
+  getReference<Entity extends object, PK extends keyof Entity>(entityName: EntityName<Entity>, id: Primary<Entity>, options: Omit<GetReferenceOptions, 'wrapped'> & { wrapped: true }): Ref<Entity, PK>;
 
   /**
    * Gets a reference to the entity identified by the given type and identifier without actually loading it, if the entity is not yet loaded