+++ tests/features/embeddables/GH2717.test.ts
@@ -1,4 +1,4 @@
-import { Collection, Embeddable, Embedded, Entity, IdentifiedReference, LoadStrategy, ManyToOne, MikroORM, OneToMany, PrimaryKey, Property } from '@mikro-orm/core';
+import { Collection, Embeddable, Embedded, Entity, Ref, LoadStrategy, ManyToOne, MikroORM, OneToMany, PrimaryKey, Property } from '@mikro-orm/core';
 import { SqliteDriver } from '@mikro-orm/sqlite';
 
 @Entity()
@@ -8,7 +8,7 @@ export class Cat {
   name!: string;
 
   @ManyToOne(() => User, { wrappedReference: true })
-  user!: IdentifiedReference<User>;
+  user!: Ref<User>;
 
 }